services/identity-service/pyproject.toml
[project]
name = "identity-service"
version = "0.1.0"
description = "Identity and account management service"
authors = [{ name = "Platform Team", email = "dev@example.com" }]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8",
    "psycopg[binary]>=3.2",
    "psycopg-pool>=3.2",
    "PyJWT>=2.9",
    "email-validator>=2.2.0",
    "redis>=5.0"
]

[project.optional-dependencies]
dev = [
    "black>=24.8.0",
    "ruff>=0.6.0",
    "pytest>=8.3.0",
    "httpx>=0.27.0",
    "fakeredis>=2.23"
]

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = [
    "app",
    "pyproject.toml",
    "README.md",
    "tests"
]
